added getUser endpoint. protected admin controllers

+++ 14-RESTAPI/server/controllers/adminController.ts
@@ -16,7 +16,7 @@ const newPost: RequestHandler = async (req, res, next) => {
       return;
     }
 
-    const post = new Post({ title, content, author: '67768c6bfa4804119c44db20' });
+    const post = new Post({ title, content, author: req.user });
     if (image) post.imgURL = image.path;
     await post.save();
     res.status(201).json(post);
@@ -39,7 +39,7 @@ const editPost: RequestHandler = async (req, res, next) => {
       return;
     }
 
-    const post = await Post.findById(postId);
+    const post = await Post.findOne({ _id: postId, author: req.user })
     if (!post) {
       res.status(404).json({ message: 'Post not found.' });
       return;
@@ -66,12 +66,12 @@ const editPost: RequestHandler = async (req, res, next) => {
 const deletePost: RequestHandler = async (req, res, next) => {
   try {
     const { postId: _id } = req.params;
-    const post = await Post.findOne({ _id });
+    const post = await Post.findOne({ _id, author: req.user })
     if (post) {
       if (post.imgURL) {
         unlink(post.imgURL, (error) => error && errorMsg({ error, where: 'deletePost' }));
       }
-      await Post.deleteOne({ _id });
+      await Post.deleteOne({ _id, author: req.user });
       res.status(200).json(null); // truthy objects cause errors as they do not match Models
     }
   } catch (error) {
